feat(plugin): add chat channel change message

+++ XIVChatCommon/Message/Message.cs
@@ -565,7 +565,7 @@ namespace XIVChatCommon.Message {
         OtherPet = 2048,
     }
 
-    public enum InputChannel : byte {
+    public enum InputChannel : uint {
         Tell = 0,
         Say = 1,
         Party = 2,
